\name{xtal-package}
\alias{xtal-package}
\alias{xtal}
\docType{package}
\title{
Crystallization Tool
}
\description{
This is the tool set for crystallographer to  design and analyze crystallization experiments, especially for ribosome from Mycobacterium tuberculosis.
}
\details{
\tabular{ll}{
Package: \tab xtal\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-12-28\cr
License: \tab GPL-2|GPL-3\cr
Depends: \tab methods,graphics, grDevices, stats, utils\cr
}
}
\author{
Qingan Sun, Xiaojun Li

Maintainer: Qingan Sun <quinsun@gmail.com>
}
\keyword{ package }
